[MAT-130] refactor: User&Match dto 패키지 구조 변경
