+++ src/main/java/com/matchus/domains/match/converter/MatchConverter.java
@@ -4,8 +4,8 @@ import com.matchus.domains.common.AgeGroup;
 import com.matchus.domains.common.Period;
 import com.matchus.domains.location.domain.Location;
 import com.matchus.domains.match.domain.Match;
-import com.matchus.domains.match.dto.MatchCreateRequest;
-import com.matchus.domains.match.dto.MatchInfoResponse;
+import com.matchus.domains.match.dto.request.MatchCreateRequest;
+import com.matchus.domains.match.dto.response.MatchInfoResponse;
 import com.matchus.domains.sports.domain.Sports;
 import com.matchus.domains.team.domain.Team;
 import org.springframework.stereotype.Component;
