+++ src/main/java/com/matchus/domains/match/dto/response/MatchInfoResponse.java
@@ -1,4 +1,4 @@
-package com.matchus.domains.match.dto;
+package com.matchus.domains.match.dto.response;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
