+++ src/main/java/com/matchus/domains/match/service/MatchService.java
@@ -6,10 +6,10 @@ import com.matchus.domains.match.converter.MatchConverter;
 import com.matchus.domains.match.domain.Match;
 import com.matchus.domains.match.domain.TeamWaiting;
 import com.matchus.domains.match.domain.WaitingType;
-import com.matchus.domains.match.dto.MatchCreateRequest;
-import com.matchus.domains.match.dto.MatchCreateResponse;
-import com.matchus.domains.match.dto.MatchInfoResponse;
-import com.matchus.domains.match.dto.MatchMember;
+import com.matchus.domains.match.dto.request.MatchCreateRequest;
+import com.matchus.domains.match.dto.response.MatchCreateResponse;
+import com.matchus.domains.match.dto.response.MatchInfoResponse;
+import com.matchus.domains.match.dto.response.MatchMember;
 import com.matchus.domains.match.exception.MatchNotFoundException;
 import com.matchus.domains.match.repository.MatchRepository;
 import com.matchus.domains.sports.domain.Sports;
