+++ src/main/java/com/matchus/domains/user/service/UserService.java
@@ -7,9 +7,9 @@ import com.matchus.domains.team.service.TeamUserService;
 import com.matchus.domains.user.converter.UserConverter;
 import com.matchus.domains.user.domain.Grouping;
 import com.matchus.domains.user.domain.User;
-import com.matchus.domains.user.dto.LoginRequest;
-import com.matchus.domains.user.dto.LoginResponse;
-import com.matchus.domains.user.dto.SignUpRequest;
+import com.matchus.domains.user.dto.request.LoginRequest;
+import com.matchus.domains.user.dto.request.SignUpRequest;
+import com.matchus.domains.user.dto.response.LoginResponse;
 import com.matchus.domains.user.exception.UserNotFoundException;
 import com.matchus.domains.user.repository.UserRepository;
 import com.matchus.global.error.ErrorCode;
